% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{census_divisions_ca21}
\alias{census_divisions_ca21}
\title{Census divisions (CD) boundaries and associated data (2021).}
\format{
A simple feature class (sf) polygon object containing 293 rows and 7 variables; each row represents a unique CD with associated geo-referenced geometry. Group of neighbouring municipalities joined together for the purposes of regional planning and managing common services (such as police or ambulance services). These groupings are established under laws in effect in certain provinces of Canada. Census division (CD) is the general term for provincially legislated areas (such as county, municipalité régionale de comté (MRC) and regional district) or their equivalents. Census divisions are intermediate geographic areas between the province/territory level and the municipality (census subdivision).
\describe{
\item{CDUID}{Uniquely identifies a census division (composed of the 2-digit province or territory unique identifier followed by the 2-digit census division code).}
\item{DGUID}{Dissemination Geography Unique Identifier.}
\item{CDNAME}{Census division name.}
\item{CDTYPE}{Census division type.}
\item{LANDAREA}{Land area of geographic area, in square kilometres (km2)}
\item{PRUID}{Uniquely identifies a province or territory.}
\item{geometry}{The sfc polygon geometry (cartographic boundaries).}
}
}
\source{
"The 2021 census dissemination areas files as created by the Canadian Census available \href{https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/index2021-eng.cfm?year=21}{here} accessed May 6th 2024. All variable definitions are based on the definitions included in the census year 2021 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2021 Census. Statistics Canada Catalogue no. 92-160-G.).
}
\usage{
data(census_divisions_ca21)
}
\description{
This object contains the census divisions (CD) for all Canadian provinces and territories (2021).
}
\keyword{census}
\keyword{divisions}
