% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{census_tracts_ca21}
\alias{census_tracts_ca21}
\title{Census tract (CT) boundaries and associated data (2021).}
\format{
A simple feature class (sf) polygon object containing 6,247 rows and 10 variables; each row represents a unique CT with associated geo-referenced geometry.
\describe{
\item{CTUID}{Uniquely identifies a census tract. Each CT is assigned a seven‑character numeric "name" (including leading zeros, decimal point and trailing zeros).}
\item{DGUID}{Census Tract Geography Unique Identifier.}
\item{CTNAME}{Census tract name.}
\item{LANDAREA}{Land area of geographic area, in square kilometres (km2)}
\item{PRUID}{Uniquely identifies a province or territory.}
\item{CMAUID}{Census Metropolitan Area or Census Agglomeration unique identifier.}
\item{CMANAME}{Census Metropolitan Area or Census Agglomeration name.}
\item{PCD}{Uniquely identifies the census division.}
\item{CDNAME}{Census division name.}
\item{geometry}{The sfc polygon geometry (cartographic boundaries).}
}
}
\source{
"The 2021 census tracts areas files as created by the Canadian Census available \href{https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/index2021-eng.cfm?year=21}{here} accessed May 6th 2024. All variable definitions are based on the definitions included in the census year 2021 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2021 Census. Statistics Canada Catalogue no. 92-160-G.).
}
\usage{
data(census_tracts_ca21)
}
\description{
This object contains the census tracts (CT) for all census metropolitan areas and census agglomerations (2021).
CTs are small, relatively stable geographic areas that usually have a population of fewer than 7,500 persons,
based on data from the previous Census of Population Program. They are located in census metropolitan areas (CMAs)
and in census agglomerations (CAs) that had a core population of 50,000 or more in the previous census.
}
\keyword{census}
\keyword{tract}
