% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{cma_education_impedance_functions}
\alias{cma_education_impedance_functions}
\title{Impedance Functions by Education Level for Census Metropolitan Areas}
\format{
A data frame with rows representing unique parameterized
impedance functions for each combination of CMA, transportation mode,
and education level.
\describe{
\item{CMA}{Census metropolitan area code (3-digit).}
\item{CMANAME}{CMA or CA name.}
\item{PR}{Province code (2-digit).}
\item{PRNAME}{Province or territory name.}
\item{Education}{Education level. One of: 'High school or no certificate', 'College or apprenticeship certificate', or 'University certificate or higher'.}
\item{PwMode_label}{Transportation mode. One of: 'Bike', 'Car-motorized', 'Public transit', or 'Walk'.}
\item{Distribution}{Best-fitting probability distribution. One of: 'Exponential', 'Normal', 'Log-normal', 'Gamma', or 'Uniform'.}
\item{est_1}{First parameter estimate for the specified distribution.}
\item{est_2}{Second parameter estimate for the specified distribution.}
}
}
\source{
Statistics Canada. (2021). Census of Population.
\url{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}
Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
This aggregated dataset was produced from confidential microdata accessed at the
Research Data Centre (RDC) at McMaster University and was vetted and approved
for release in accordance with disclosure control requirements.
Accessed December 19, 2024.
}
\usage{
data(cma_education_impedance_functions)
}
\description{
Impedance functions fitted to commuting duration distributions for
Canadian Census Metropolitan Areas (CMAs) and Census Agglomerations (CAs),
categorized by education level. Functions were estimated using the
\code{fitdistrplus} package based on PwDUR data from the 2021 Census of Population.
The best-fit distribution was selected using the lowest AIC.
}
\keyword{area}
\keyword{census}
\keyword{education}
\keyword{functions}
\keyword{impedance}
\keyword{metropolitan}
\keyword{time}
\keyword{travel}
