% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{cma_keys}
\alias{cma_keys}
\title{Unique Census Metropolitan Area (CMA) Keys}
\format{
A data frame with rows representing unique CMAs/CAs.
All geographic identifiers are stored as factors.
\describe{
\item{CMA}{Census Metropolitan Area or Census Agglomeration code (3-digit).}
\item{CMANAME}{CMA or CA name.}
\item{CMACATEG}{CMA category. One of: 'Census Metropolitan Area or Census Agglomeration', 'Strong metropolitan influenced zone', 'Moderate metropolitan influenced zone', 'Weak metropolitan influenced zone', 'No metropolitan influenced zone', or 'Territories (outside CAs)'.}
\item{PR}{Province code (2-digit).}
\item{PRNAME}{Province or territory name.}
}
}
\source{
Obtained from \code{\link{census_geographical_keys}}.
}
\usage{
data(cma_keys)
}
\description{
A version of the geographic keys "census_geographical_keys" containing one row per
unique Census Metropolitan Area (CMA) or Census Agglomeration (CA).
}
\keyword{cma}
\keyword{geography}
\keyword{key}
