% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{package}
\name{CommuteCA-package}
\alias{CommuteCA}
\alias{CommuteCA-package}
\title{CommuteCA: Datasets and Methods for Job Accessibility in Canada}
\description{
Provides datasets and standardized methods for job accessibility analysis using the
\href{https://www12.statcan.gc.ca/census-recensement/index-eng.cfm}{\emph{2021 Census of Population}} of Statistics Canada. Includes synthetic test data,
aggregated land-use datasets, and calibrated impedance functions for commuting
by transportation mode. Created with support from the \href{https://rdc.mcmaster.ca/}{\emph{Research Data Centre} at \emph{McMaster University}},
the \href{https://scds.ca/}{\emph{Sherman Centre for Digital Scholarship}}, and \href{https://mobilizingjustice.ca/}{\emph{Mobilizing Justice}}.
}
\section{Data Attribution}{

Data from Statistics Canada (2021 Census of Population) is included under
the Statistics Canada Open Licence and is reproduced and distributed on
an "as is" basis with the permission of Statistics Canada.
Source: Statistics Canada, 2021 Census of Population.
}

\references{
\url{https://github.com/dias-bruno/CommuteCA}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/CommuteCA}
  \item \url{https://dias-bruno.github.io/CommuteCA/}
  \item Report bugs at \url{https://github.com/dias-bruno/CommuteCA/issues}
}

}
\author{
Bruno Dos Santos, School of Earth, Environment and Society, McMaster University \email{dossanb@mcmaster.ca}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
