% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{ct_keys}
\alias{ct_keys}
\title{Unique Census Tract (CT) Keys}
\format{
A data frame with rows representing unique CTs.
All geographic identifiers are stored as factors.
\describe{
\item{CTUID}{Census Tract unique identifier (7-digit code).}
\item{CTNAME}{Census Tract name.}
\item{PCSD}{Census Subdivision unique identifier (7-digit code).}
\item{PCSDNAME}{Census Subdivision name.}
\item{PCD}{Census Division code (4-digit).}
\item{PCDNAME}{Census Division name.}
\item{CMA}{Census Metropolitan Area or Census Agglomeration code (3-digit).}
\item{CMANAME}{CMA or CA name.}
\item{CMACATEG}{CMA category. One of: 'Census Metropolitan Area or Census Agglomeration', 'Strong metropolitan influenced zone', 'Moderate metropolitan influenced zone', 'Weak metropolitan influenced zone', 'No metropolitan influenced zone', or 'Territories (outside CAs)'.}
\item{PR}{Province code (2-digit).}
\item{PRNAME}{Province or territory name.}
}
}
\source{
Obtained from \code{\link{census_geographical_keys}}.
}
\usage{
data(ct_keys)
}
\description{
A version of the geographic keys "census_geographical_keys" containing one row per
unique Census Tract (CT).
}
\keyword{census}
\keyword{ct}
\keyword{geography}
\keyword{key}
\keyword{tract}
