% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{dissemination_areas_ca21}
\alias{dissemination_areas_ca21}
\title{Dissemination areas (DA) boundaries and associated data (2021).}
\format{
A simple feature class (sf) polygon object containing 57,932 rows and 8 variables; each row represents a unique dissemination area (DA) with associated geo-referenced geometry. A dissemination area (DA) is a small, relatively stable geographic unit composed of one or more adjacent dissemination blocks. It is the smallest standard geographic area for which all census data are disseminated. DAs cover all Canada.
\describe{
\item{DAUID}{Uniquely identifies a dissemination area (composed of the 2-digit province/territory unique identifier followed by the 2-digit census division code and the 4-digit dissemination area code).}
\item{DGUID}{Dissemination Geography Unique Identifier.}
\item{LANDAREA}{Land area of geographic area, in square kilometres (km2)}
\item{PRUID}{Uniquely identifies a province or territory.}
\item{PCD}{Uniquely identifies the census division.}
\item{CMAUID}{Uniquely identifies the census metropolitan area or census agglomeration (if any).}
\item{CMANAME}{Name of the the census metropolitan area or census agglomeration (if any).}
\item{geometry}{The sfc polygon geometry (cartographic boundaries).}
}
}
\source{
"The 2021 census dissemination areas files as created by the Canadian Census available \href{https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/index2021-eng.cfm?year=21}{here} accessed May 6th 2024. All variable definitions are based on the definitions included in the census year 2021 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2021 Census. Statistics Canada Catalogue no. 92-160-G.).
}
\usage{
data(dissemination_areas_ca21)
}
\description{
This object contains the dissemination areas (DA) for all Canadian provinces and territories (2021).
}
\keyword{areas}
\keyword{dissemination}
