% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{HT_Toronto_general_DA}
\alias{HT_Toronto_general_DA}
\title{Toronto job accessibility (Hansen type)}
\format{
A data frame containing 3,727 rows and 2 variables; each row represents a unique DA.
\describe{
\item{PRCDDA}{Uniquely identifies a dissemination area (composed of the 2-digit province/territory unique identifier followed by the 2-digit census division code and the 4-digit dissemination area code).}
\item{HT_i}{DA job accessibility (Hansen Type).}
}
}
\source{
"The 2021 census tracts areas files as created by the Canadian Census available \href{https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/index2021-eng.cfm?year=21}{here} accessed May 6th 2024. All variable definitions are based on the definitions included in the census year 2021 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2021 Census. Statistics Canada Catalogue no. 92-160-G.).
}
\usage{
data(HT_Toronto_general_DA)
}
\description{
This object contains the value of job accessibility for the City of Toronto, at the level of Dissemination Areas (DAs).
The measures were obtained applying the Hansen type metric, using the Census of Population (2021) as data source.
}
\keyword{Hansen}
\keyword{Toronto}
\keyword{accessibility}
\keyword{area}
\keyword{dissemination}
