% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{jobs_CT_general}
\alias{jobs_CT_general}
\title{Job Opportunities by Census Tract}
\format{
A data frame with rows representing unique Census Tracts.
\describe{
\item{CTUID}{Census tract unique identifier (7-digit code).}
\item{CTNAME}{Census tract name.}
\item{PCD}{Census division code (4-digit).}
\item{CMA}{Census metropolitan area code (3-digit).}
\item{PR}{Province code (2-digit).}
\item{jobs_rounded}{Total job opportunities in the CT, rounded to the nearest multiple of 5.}
}
}
\source{
Statistics Canada. (2021). Census of Population.
\url{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}
Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
This aggregated dataset was produced from confidential microdata accessed at the
Research Data Centre (RDC) at McMaster University and was vetted and approved
for release in accordance with disclosure control requirements.
Accessed December 19, 2024.
}
\usage{
data(jobs_CT_general)
}
\description{
Total job opportunities (employment locations) for each Census Tract (CT)
in Canada, obtained from the 2021 Census of Population. Job counts are
calculated by summing the weights (CompW1) of all workers whose place of
work is within the CT. This dataset has been vetted and approved for
release by the Research Data Centre (RDC) at McMaster University.
}
\keyword{census}
\keyword{employment}
\keyword{jobs}
\keyword{land}
\keyword{tract}
\keyword{use}
