% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{pcd_keys}
\alias{pcd_keys}
\title{Unique Census Division (CD) Keys}
\format{
A data frame with rows representing unique CDs.
All geographic identifiers are stored as factors.
\describe{
\item{PCD}{Census Division code (4-digit: 2-digit province + 2-digit division).}
\item{PCDNAME}{Census Division name.}
\item{CMA}{Census Metropolitan Area or Census Agglomeration code (3-digit).}
\item{CMANAME}{CMA or CA name.}
\item{CMACATEG}{CMA category. One of: 'Census Metropolitan Area or Census Agglomeration', 'Strong metropolitan influenced zone', 'Moderate metropolitan influenced zone', 'Weak metropolitan influenced zone', 'No metropolitan influenced zone', or 'Territories (outside CAs)'.}
\item{PR}{Province code (2-digit).}
\item{PRNAME}{Province or territory name.}
}
}
\source{
Obtained from \code{\link{census_geographical_keys}}.
}
\usage{
data(pcd_keys)
}
\description{
A version of the geographic keys "census_geographical_keys" containing one row per
unique Census Division (CD).
}
\keyword{census}
\keyword{division}
\keyword{geography}
\keyword{key}
\keyword{pcd}
