% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{pcd_max_travel_times}
\alias{pcd_max_travel_times}
\title{Maximum Travel Times by Census Division}
\format{
A data frame with rows representing each combination of
Census Division and transportation mode.
\describe{
\item{PCD}{Census division code (4-digit: 2-digit province + 2-digit division).}
\item{PCDNAME}{Census division name.}
\item{PR}{Province code (2-digit).}
\item{PRNAME}{Province or territory name.}
\item{PwMode_label}{Transportation mode. One of: 'Bike', 'Car-motorized', 'Public transit', or 'Walk'.}
\item{Maximum_travel_time}{Maximum commuting duration (in minutes) for the given CD and mode.}
}
}
\source{
Statistics Canada. (2021). Census of Population.
\url{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}
Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
This aggregated dataset was produced from confidential microdata accessed at the
Research Data Centre (RDC) at McMaster University and was vetted and approved
for release in accordance with disclosure control requirements.
Accessed December 19, 2024.
}
\usage{
data(pcd_max_travel_times)
}
\description{
Maximum travel times for commuting to work for all Census Divisions (CDs)
in Canada, obtained from the 2021 Census of Population. Travel times are
calculated separately for each transportation mode.
}
\keyword{census}
\keyword{division}
\keyword{maximum}
\keyword{time}
\keyword{travel}
