% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{data}
\name{census_test_toronto}
\alias{census_test_toronto}
\title{2021 Census of Population Test Data for the City of Toronto}
\format{
A data frame object containing 56,650 rows and 26 variables; each row represents a unique (fake) respondent.
\describe{
\item{Frame_ID}{Unique identifier of the (fake) respondent.}
\item{PRCDDA}{Refers to the dissemination area (DA) of current residence}
\item{Pr}{Refers to the province or territory of current residence.}
\item{CMA}{Census metropolitan area or census agglomeration of current residence.}
\item{PCD}{Census division of current residence.}
\item{CompW1}{Weight for the households and dwellings universes.}
\item{LBR_FORC}{This variable refers to whether a person was employed, unemployed or not in the labour force.}
\item{CfInc}{Total income of census family (sum of the total incomes of all members of that family).}
\item{CFCNT}{Census family size (numbers of persons).}
\item{CF_PnCF_NumEarners}{Number of earners in census family.}
\item{PWDA}{Place of work dissemination area.}
\item{PWPR}{Place of work province.}
\item{PWCD}{Census metropolitan area or census agglomeration of place of work.}
\item{PWCMA}{Place of work census division.}
\item{PWDUR}{Commuting duration, it refers to the length of time, in minutes, usually required by a person to travel to their place of work.}
\item{PWDist}{Distance (straight line) from home to work.}
\item{PwMode}{Main mode of commuting' refers to the main mode of transportation a person uses to travel to their place of work.}
\item{AGE}{Age of the (fake) respondent.}
\item{GENDER_BINARY}{Gender of (fake) person (binary).}
\item{LOLICOA}{Low-income status based on LICO-AT (after tax).}
\item{LoLIMA}{Low-income status based on LIM-AT (after tax).}
\item{ACTLMFLG}{Activities of daily living: Flag.}
\item{CFAMSTSIMPLE}{Type of (fake) family.}
\item{IMMCAT5}{Immigration: Admission category.}
\item{PerImmA}{Immigration: Immigrant status and period of immigration (by five year group).}
\item{DVisMin}{If the (fake) person is part of a visible minority group.}
}
}
\usage{
data(census_test_toronto)
}
\description{
This object contains test data (fake observations created to represent the original version) of the Census of Population for the City of Toronto.
}
\keyword{census}
\keyword{data}
\keyword{test}
\keyword{toronto}
