% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA-package.R
\docType{package}
\name{CommuteCA-package}
\alias{CommuteCA}
\alias{CommuteCA-package}
\title{CommuteCA: an R package to develop standardized methods for transport analysis in research}
\description{
The main objective of the \emph{CommuteCA} package is to develop standardized methods for
accessibility analysis in research, especially for studies using Statistics Canada surveys.
Among the available surveys, we focused our efforts on the \href{https://www12.statcan.gc.ca/census-recensement/index-eng.cfm}{\emph{2021 Census of Population}},
which contain valuable variables for transportation research.
This package was created in conjunction with the office of the
\href{https://rdc.mcmaster.ca/}{\emph{Research Data Center} at \emph{McMaster University}},
the \href{https://scds.ca/}{\emph{Sherman Centre for Digital Scholarship}} and
the \href{https://mobilizingjustice.ca/}{\emph{Mobilizing Justice}}.
}
\references{
\url{https://github.com/dias-bruno/CommuteCA}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/CommuteCA}
  \item \url{https://dias-bruno.github.io/CommuteCA/}
  \item Report bugs at \url{https://github.com/dias-bruno/CommuteCA/issues}
}

}
\author{
Bruno Dos Santos, School of Earth, Environment and Society, McMaster University \email{dossanb@mcmaster.ca}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
